scripts/requirements.txt
azure-search-documents==11.6.0b2
langchain==0.1.11
azure-ai-formrecognizer==3.3.2
openai==1.13.3
azure-identity==1.15.0